+++ commons/src/main/java/com/karuslabs/commons/command/annotations/old/assembler/CommandAssembler.java
@@ -23,7 +23,8 @@
  */
 package com.karuslabs.commons.command.annotations.old.assembler;
 
-import com.karuslabs.commons.command.annotations.*;
+import com.karuslabs.commons.command.aot.annotations.Argument;
+import com.karuslabs.commons.command.aot.annotations.Literal;
 import com.karuslabs.commons.util.collection.TokenMap;
 
 import com.mojang.brigadier.Command;