+++ commons/src/main/java/com/karuslabs/commons/command/aot/lexers/ArgumentLexer.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2019 Karus Labs.
+ * Copyright 2020 Karus Labs.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,19 +21,29 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.karuslabs.commons.command.annotations;
-
-import java.lang.annotation.*;
-
-import static java.lang.annotation.ElementType.*;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
+package com.karuslabs.commons.command.aot.lexers;
 
 
-@Documented
-@Retention(RUNTIME)
-@Target(METHOD)
-public @interface Argument {
-    
-    String[] value();
-    
+public class ArgumentLexer implements Lexer {
+
+    @Override
+    public void lex(Visitor visitor, String context, String raw) {
+        if (!raw.startsWith("<") || !raw.endsWith(">")) {
+            visitor.error("Invalid argument syntax: '" + raw + "' in '" + context + "', arguments must be enclosed by '<' and '>'");
+        }
+        
+        if (raw.contains("\\|")) {
+            visitor.error("Invalid argument syntax: '" + raw + "' in '" + context + "', arguments cannot contain '|'");
+            return;
+        }
+        
+        var argument = raw.substring(1, raw.length() - 1);
+        if (argument.isBlank()) {
+            visitor.error("Invalid argument name in '" + context + "', argument names cannot be blank");
+            
+        } else {
+            visitor.argument(context, argument);
+        }
+    }
+
 }