+++ commons/src/main/java/com/karuslabs/commons/command/aot/lints/Lint.java
@@ -21,24 +21,36 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.karuslabs.commons.command.annotations.lint;
+package com.karuslabs.commons.command.aot.lints;
 
-import java.util.Map;
+import com.karuslabs.commons.command.aot.Node;
+import com.karuslabs.commons.command.aot.lexers.Lexer;
+
+import java.util.*;
 import javax.annotation.processing.Messager;
 import javax.lang.model.element.Element;
 import javax.lang.model.util.*;
 
+import org.checkerframework.checker.nullness.qual.Nullable;
 
-public abstract class Lint {
+import static javax.tools.Diagnostic.Kind.ERROR;
+
+
+public abstract class Lint implements Lexer.Visitor {
     
-    protected Map<String, Entry> commands;
+    protected Lexer lexer;
+    protected Map<Element, Set<Node>> scopes;
+    protected Node root;
+    protected @Nullable Element element;
     protected Messager messager;
     protected Elements elements;
     protected Types types;
     
     
-    public Lint(Map<String, Entry> commands, Messager messager, Elements elements, Types types) {
-        this.commands = commands;
+    public Lint(Lexer lexer, Map<Element, Set<Node>> scopes, Node root, Messager messager, Elements elements, Types types) {
+        this.lexer = lexer;
+        this.scopes = scopes;
+        this.root = root;
         this.messager = messager;
         this.elements = elements;
         this.types = types;
@@ -46,5 +58,21 @@ public abstract class Lint {
     
     
     public abstract void lint(Element element);
+    
+    
+    @Override
+    public void error(String message) {
+        messager.printMessage(ERROR, message, element);
+    }
+    
+    
+    protected Set<Node> scope(Element element) {
+        var scope = scopes.get(element);
+        if (scope == null) {
+            scopes.put(element, scope = new HashSet<>());
+        }
+        
+        return scope;
+    }
     
 }